Adattati i file all'aggiunta di ResponseHandler

+++ apiPB/Controllers/RegOreController.cs
@@ -13,14 +13,14 @@ namespace apiPB.Controllers
     [ApiController]
     public class RegOreController : Controller
     {
-        private readonly ILogService _logService;
+        private readonly IResponseHandler _responseHandler;
         private readonly IRegOreRequestService _regOreRequestService;
-        private readonly bool _logIsActive;
-        public RegOreController(ILogService logService, IRegOreRequestService regOreRequestService)
+        private readonly bool _isLogActive;
+        public RegOreController(IResponseHandler responseHandler, IRegOreRequestService regOreRequestService)
         {
-            _logService = logService;
+            _responseHandler = responseHandler;
             _regOreRequestService = regOreRequestService;
-            _logIsActive = false;
+            _isLogActive = false;
         }
 
         [HttpGet("get_all")]
@@ -29,22 +29,13 @@ namespace apiPB.Controllers
         /// </summary>
         /// <response code="200">Ritorna tutte le informazioni della vista A3_app_reg_ore</response>
         /// <response code="404">Non trovato</response>
-        public IActionResult getAllRegOre()
-        {
-            string requestPath = $"{HttpContext.Request.Method} {HttpContext.Request.Path.Value?.TrimStart('/') ?? string.Empty}";
-                
+        public IActionResult GetAllRegOre()
+        {   
             var a3AppRegOreDto = _regOreRequestService.GetAppRegOre().ToList();
 
-            if (a3AppRegOreDto.IsNullOrEmpty())
-            {
-                _logService.AppendMessageToLog(requestPath, NotFound().StatusCode, "Not Found", _logIsActive);
+            if (a3AppRegOreDto.IsNullOrEmpty()) return _responseHandler.HandleNotFound(HttpContext, _isLogActive);
 
-                return NotFound();
-            }
-
-            _logService.AppendMessageAndListToLog(requestPath, Ok().StatusCode, "OK", a3AppRegOreDto, _logIsActive);
-
-            return Ok(a3AppRegOreDto);
+            return _responseHandler.HandleOkAndList(HttpContext, a3AppRegOreDto, _isLogActive);
         }
 
         [HttpPost("post_reg_ore")]
@@ -54,28 +45,15 @@ namespace apiPB.Controllers
         /// <param name="IEnumerable<a3AppRegOreRequestDto>">Collezione contenente i parametri di ricerca</param>
         /// <response code="201">Crea delle entry nel database</response>
         /// <response code="404">Non trovato</response>
-        public IActionResult PostRegOreList([FromBody] IEnumerable<RegOreRequestDto> a3AppRegOreRequestDto)
+        public IActionResult PostRegOreList([FromBody] IEnumerable<RegOreRequestDto>? a3AppRegOreRequestDto)
         {
-            if (a3AppRegOreRequestDto.IsNullOrEmpty())
-            {
-                return BadRequest("La richiesta non può essere vuota.");
-            }
-            
-            string requestPath = $"{HttpContext.Request.Method} {HttpContext.Request.Path.Value?.TrimStart('/') ?? string.Empty}";
+            if (a3AppRegOreRequestDto == null || !a3AppRegOreRequestDto.Any()) return _responseHandler.HandleBadRequest(HttpContext, _isLogActive);
 
             var a3AppRegOreDto = _regOreRequestService.PostAppRegOre(a3AppRegOreRequestDto).ToList();
 
-            if (a3AppRegOreDto.IsNullOrEmpty())
-            {
-                _logService.AppendMessageToLog(requestPath, NotFound().StatusCode, "Not Found", _logIsActive);
+            if (a3AppRegOreDto.IsNullOrEmpty()) return _responseHandler.HandleNotFound(HttpContext, _isLogActive);
 
-                return NotFound();
-            }
-
-            var createdResult = CreatedAtAction(nameof(PostRegOreList), a3AppRegOreDto);
-            _logService.AppendMessageAndListToLog(requestPath, createdResult.StatusCode, "Created", a3AppRegOreDto, _logIsActive);
-
-            return CreatedAtAction(nameof(PostRegOreList), a3AppRegOreDto);
+            return _responseHandler.HandleCreated(HttpContext, a3AppRegOreDto, _isLogActive);
         }
 
         [HttpPost("view_ore")]
@@ -85,22 +63,15 @@ namespace apiPB.Controllers
         /// <param name="A3AppViewOreRequestDto">Oggetto contenente i parametri di ricerca</param>
         /// <response code="200">Ritorna tutte le informazioni della vista A3_app_reg_ore filtrate</response>
         /// <response code="404">Non trovato</response>
-        public IActionResult GetA3AppRegOre([FromBody] ViewOreRequestDto a3AppViewOreRequestDto)
+        public IActionResult GetA3AppRegOre([FromBody] ViewOreRequestDto? a3AppViewOreRequestDto)
         {
-            string requestPath = $"{HttpContext.Request.Method} {HttpContext.Request.Path.Value?.TrimStart('/') ?? string.Empty}";
+            if(a3AppViewOreRequestDto == null) return _responseHandler.HandleBadRequest(HttpContext, _isLogActive);
 
             var a3AppRegOreDto = _regOreRequestService.GetAppViewOre(a3AppViewOreRequestDto).ToList();
 
-            if (a3AppRegOreDto.IsNullOrEmpty())
-            {
-                _logService.AppendMessageToLog(requestPath, NotFound().StatusCode, "Not Found", _logIsActive);
+            if (a3AppRegOreDto.IsNullOrEmpty()) return _responseHandler.HandleNotFound(HttpContext, _isLogActive);
 
-                return NotFound();
-            }
-
-            _logService.AppendMessageAndListToLog(requestPath, Ok().StatusCode, "OK", a3AppRegOreDto, _logIsActive);
-
-            return Ok(a3AppRegOreDto);
+            return _responseHandler.HandleOkAndList(HttpContext, a3AppRegOreDto, _isLogActive);
         }
 
         [HttpPut("view_ore/edit_working_time")]
@@ -110,22 +81,15 @@ namespace apiPB.Controllers
         /// <param name="A3AppViewOrePutRequestDto">Oggetto contenente i parametri di ricerca</param>
         /// <response code="200">Ritorna il record modificato della tabella A3_app_reg_ore</response>
         /// <response code="404">Non trovato</response>
-        public IActionResult PutA3AppRegOre([FromBody] ViewOrePutRequestDto a3AppViewOrePutRequestDto)
+        public IActionResult PutA3AppRegOre([FromBody] ViewOrePutRequestDto? a3AppViewOrePutRequestDto)
         {
-            string requestPath = $"{HttpContext.Request.Method} {HttpContext.Request.Path.Value?.TrimStart('/') ?? string.Empty}";
+            if(a3AppViewOrePutRequestDto == null) return _responseHandler.HandleBadRequest(HttpContext, _isLogActive);
 
             var a3AppRegOreDto = _regOreRequestService.PutAppViewOre(a3AppViewOrePutRequestDto);
 
-            if (a3AppRegOreDto == null)
-            {
-                _logService.AppendMessageToLog(requestPath, NotFound().StatusCode, "Not Found", _logIsActive);
+            if (a3AppRegOreDto == null) return _responseHandler.HandleNotFound(HttpContext, _isLogActive);
 
-                return NotFound();
-            }
-
-            _logService.AppendMessageAndItemToLog(requestPath, Ok().StatusCode, "OK", a3AppRegOreDto, _logIsActive);
-
-            return Ok(a3AppRegOreDto);
+            return _responseHandler.HandleOkAndItem(HttpContext, a3AppRegOreDto, _isLogActive);
         }
 
         [HttpDelete("view_ore/delete_reg_ore_id")]
@@ -135,22 +99,15 @@ namespace apiPB.Controllers
         /// <param name="A3AppDeleteRequestDto">Oggetto contenente i parametri di ricerca</param>
         /// <response code="200">Ritorna il record eliminato della tabella A3_app_reg_ore</response>
         /// <response code="404">Non trovato</response>
-        public IActionResult DeleteRegOreId([FromBody] ViewOreDeleteRequestDto a3AppDeleteRequestDto)
+        public IActionResult DeleteRegOreId([FromBody] ViewOreDeleteRequestDto? a3AppDeleteRequestDto)
         {
-            string requestPath = $"{HttpContext.Request.Method} {HttpContext.Request.Path.Value?.TrimStart('/') ?? string.Empty}";
+            if(a3AppDeleteRequestDto == null) return _responseHandler.HandleBadRequest(HttpContext, _isLogActive);
 
             var a3AppRegOreDto = _regOreRequestService.DeleteRegOreId(a3AppDeleteRequestDto);
 
-            if (a3AppRegOreDto == null)
-            {
-                _logService.AppendMessageToLog(requestPath, NotFound().StatusCode, "Not Found", _logIsActive);
+            if (a3AppRegOreDto == null) return _responseHandler.HandleNotFound(HttpContext, _isLogActive);
 
-                return NotFound();
-            }
-
-            _logService.AppendMessageAndItemToLog(requestPath, Ok().StatusCode, "Deleted", a3AppRegOreDto, _logIsActive);
-
-            return Ok(a3AppRegOreDto);
+            return _responseHandler.HandleOkAndItem(HttpContext, a3AppRegOreDto, _isLogActive);
         }
     }
 }
